Gate debug tools and the bench harness behind build tags
Release builds no longer compile KD_PPROF, the rekey threshold
overrides or the transport bench harness (-tags debug / bench,
make test-bench runs in CI). Add SPDX headers in pkg/transport,
dedup test helpers, anchor the kd/cli gitignore entries.
