+++ pkg/transport/migratable.go
@@ -1,3 +1,12 @@
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (c) 2025 KeibiSoft S.R.L.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// ABOUTME: MigratableConn is a QUIC net.Conn that migrates to a fresh local UDP
+// ABOUTME: socket (Wi-Fi to LTE) while the stream and layers above it keep working.
+
 package transport
 
 import (
