+++ pkg/transport/quiclisten.go
@@ -1,3 +1,12 @@
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (c) 2025 KeibiSoft S.R.L.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// ABOUTME: quicListener adapts a quic.Listener to net.Listener, accepting each
+// ABOUTME: connection's first stream as one net.Conn without head-of-line blocking.
+
 package transport
 
 import (
