+++ pkg/transport/secure.go
@@ -1,3 +1,14 @@
+//go:build bench
+
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (c) 2025 KeibiSoft S.R.L.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// ABOUTME: Secure runs a hybrid post-quantum handshake (ML-KEM-1024 + X25519) over
+// ABOUTME: a net.Conn and wraps it in an AEAD-framed secureConn (ChaCha20-Poly1305).
+
 package transport
 
 import (
