+++ pkg/transport/secure_kd.go
@@ -1,3 +1,14 @@
+//go:build bench
+
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (c) 2025 KeibiSoft S.R.L.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// ABOUTME: Identity and Secure run KeibiDrop's real post-quantum handshake
+// ABOUTME: (pkg/crypto, unchanged) over a QUIC stream, a drop-in for a real peer.
+
 package transport
 
 import (
