+++ pkg/transport/secure_kd_test.go
@@ -1,8 +1,17 @@
+//go:build bench
+
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (c) 2025 KeibiSoft S.R.L.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
 package transport
 
 import (
 	"bytes"
 	"context"
+	"fmt"
 	"io"
 	"net"
 	"testing"
@@ -140,19 +149,14 @@ func TestGRPCOverKDSecureQUIC(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Download: %v", err)
 	}
-	var got uint64
-	for {
-		chunk, err := stream.Recv()
-		if err == io.EOF {
-			break
+	got, err := drainDownload(stream, func(c *pb.Chunk) error {
+		if !verifyChunk(c.Data) {
+			return fmt.Errorf("corrupt chunk at offset %d", c.Offset)
 		}
-		if err != nil {
-			t.Fatalf("recv at %d: %v", got, err)
-		}
-		if !verifyChunk(chunk.Data) {
-			t.Fatalf("corrupt chunk at offset %d", chunk.Offset)
-		}
-		got += uint64(len(chunk.Data))
+		return nil
+	})
+	if err != nil {
+		t.Fatalf("recv: %v", err)
 	}
 	if got != total {
 		t.Fatalf("byte count: got %d want %d", got, total)
