+++ pkg/transport/tlsconf.go
@@ -1,3 +1,12 @@
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (c) 2025 KeibiSoft S.R.L.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// ABOUTME: serverTLSConfig and clientTLSConfig build the throwaway self-signed TLS
+// ABOUTME: config QUIC requires at the protocol level; real security is layered on top.
+
 package transport
 
 import (
